Create hoh-h20-b3lyp.gjf
Gaussian 16 input for single point calculations with B3LYP, starting from the structures optimized with MP2. For the rest of the DFAs it is the same input, you just have to change the name of the DFA in the third line. The wfx file is printed, needed to get the hydrogen bond critical points.